Implement F1 cleanup plan
