+++ apps/f1/client/src/pages/admin/AdminLoadingState.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function AdminLoadingState({ message = 'Loading admin data...' }) {
+  return <section className="loading-panel">{message}</section>;
+}
