+++ apps/f1/server/persistence/connection.js
@@ -1,0 +1,13 @@
+const Database = require('better-sqlite3');
+const path = require('path');
+
+const DB_PATH = process.env.DB_PATH || path.join(__dirname, '..', 'f1-calcutta.db');
+const db = new Database(DB_PATH);
+
+db.pragma('journal_mode = WAL');
+db.pragma('foreign_keys = ON');
+
+module.exports = {
+  db,
+  DB_PATH,
+};
