+++ apps/f1/server/routes/admin.js
@@ -1,27 +1,16 @@
 const express = require('express');
 
-const { generateInviteCode } = require('../lib/core');
-const {
-  db,
-  getActiveSeasonId,
-  getSeasonSettings,
-  updateSeasonSettings,
-  getSeasonParticipants,
-  getAuctionItems,
-  getDrivers,
-  getEventById,
-  getEventResults,
-  getEventPayoutRules,
-  getSeasonBonusRules,
-} = require('../db');
-const {
-  scoreEvent,
-  recalcSeasonBonuses,
-  upsertEventResults,
-  syncEventFromProvider,
-  syncNextEventFromProvider,
-} = require('../services/scoringService');
+const { getActiveSeasonId } = require('../db');
 const { requireAuth, requireAdmin } = require('./middleware');
+
+const auctionAdminService = require('../services/admin/auctionAdminService');
+const payoutRulesAdminService = require('../services/admin/payoutRulesAdminService');
+const resultsAdminService = require('../services/admin/resultsAdminService');
+const {
+  parseForceFlag,
+  parseIntegerParam,
+  runAndRespond,
+} = require('../services/admin/common');
 
 const router = express.Router();
 
@@ -29,280 +18,191 @@ function withAdmin(req, res, next) {
   return requireAuth(req, res, () => requireAdmin(req, res, next));
 }
 
+function parseIdFromParams(res, rawId) {
+  const id = parseIntegerParam(rawId);
+  if (id == null) {
+    res.status(400).json({ error: 'Invalid numeric id' });
+    return null;
+  }
+  return id;
+}
+
 router.get('/settings', withAdmin, (req, res) => {
-  res.json(getSeasonSettings(getActiveSeasonId()));
+  const seasonId = getActiveSeasonId();
+  return res.json(auctionAdminService.getSettings({ seasonId }));
 });
 
 router.patch('/settings', withAdmin, (req, res) => {
   const seasonId = getActiveSeasonId();
-  updateSeasonSettings(seasonId, req.body || {});
-  const settings = getSeasonSettings(seasonId);
-  req.app.get('io')?.emit('auction:status', { status: settings.auction_status });
-  return res.json({ ok: true, settings });
+  const io = req.app.get('io');
+  const result = auctionAdminService.updateSettingsForSeason({
+    seasonId,
+    patch: req.body,
+    io,
+  });
+  return runAndRespond(res, result);
 });
 
 router.post('/invite-code/regenerate', withAdmin, (req, res) => {
   const seasonId = getActiveSeasonId();
-  const code = generateInviteCode();
-  updateSeasonSettings(seasonId, { invite_code: code });
-  res.json({ ok: true, invite_code: code });
+  return runAndRespond(res, auctionAdminService.regenerateInviteCode({ seasonId }));
 });
 
 router.get('/participants', withAdmin, (req, res) => {
-  res.json(getSeasonParticipants(getActiveSeasonId()));
+  const seasonId = getActiveSeasonId();
+  return res.json(auctionAdminService.listParticipants({ seasonId }));
 });
 
 router.delete('/participants/:id', withAdmin, (req, res) => {
   const seasonId = getActiveSeasonId();
-  const participantId = parseInt(req.params.id, 10);
-  db.prepare(`
-    DELETE FROM season_participants
-    WHERE season_id = ?
-      AND participant_id = ?
-      AND participant_id IN (SELECT id FROM participants WHERE is_admin = 0)
-  `).run(seasonId, participantId);
-  res.json({ ok: true });
+  const participantId = parseIdFromParams(res, req.params.id);
+  if (participantId == null) return undefined;
+  return runAndRespond(res, auctionAdminService.removeParticipant({ seasonId, participantId }));
 });
 
 router.get('/auction/queue', withAdmin, (req, res) => {
-  res.json(getAuctionItems(getActiveSeasonId()));
+  const seasonId = getActiveSeasonId();
+  return res.json(auctionAdminService.listAuctionQueue({ seasonId }));
 });
 
 router.patch('/auction/queue', withAdmin, (req, res) => {
-  const { order } = req.body;
-  if (!Array.isArray(order)) return res.status(400).json({ error: 'order must be an array' });
-
-  const update = db.prepare(`
-    UPDATE auction_items
-    SET queue_order = ?
-    WHERE id = ? AND status = 'pending' AND season_id = ?
-  `);
-
   const seasonId = getActiveSeasonId();
-  db.transaction(() => {
-    order.forEach((item) => update.run(item.queue_order, item.id, seasonId));
-  })();
-
-  return res.json({ ok: true });
+  return runAndRespond(res, auctionAdminService.updateAuctionQueue({ seasonId, order: req.body?.order }));
 });
 
 router.post('/auction/start', withAdmin, (req, res) => {
   const seasonId = getActiveSeasonId();
-  updateSeasonSettings(seasonId, { auction_status: 'open' });
-  req.app.get('io')?.emit('auction:status', { status: 'open' });
-  res.json({ ok: true });
+  const io = req.app.get('io');
+  return runAndRespond(res, auctionAdminService.setAuctionStatus({ seasonId, status: 'open', io }));
 });
 
 router.post('/auction/pause', withAdmin, (req, res) => {
   const seasonId = getActiveSeasonId();
-  updateSeasonSettings(seasonId, { auction_status: 'paused' });
-  req.app.get('io')?.emit('auction:status', { status: 'paused' });
-  res.json({ ok: true });
+  const io = req.app.get('io');
+  return runAndRespond(res, auctionAdminService.setAuctionStatus({ seasonId, status: 'paused', io }));
 });
 
 router.post('/auction/next', withAdmin, (req, res) => {
   const seasonId = getActiveSeasonId();
   const auctionService = req.app.get('auctionService');
-  const result = auctionService.startAuction({ seasonId, driverId: req.body?.driverId });
-  if (!result.ok) return res.status(result.status || 400).json({ error: result.error });
-  return res.json({ ok: true, ...result });
+  const result = auctionAdminService.startNextAuction({
+    seasonId,
+    auctionService,
+    driverId: req.body?.driverId,
+  });
+  return runAndRespond(res, result, (payload) => ({ ok: true, ...payload }));
 });
 
 router.post('/auction/close', withAdmin, (req, res) => {
   const seasonId = getActiveSeasonId();
   const auctionService = req.app.get('auctionService');
-  const result = auctionService.closeActiveAuction({ seasonId });
-  if (!result.ok) return res.status(result.status || 400).json({ error: result.error });
-  return res.json({ ok: true });
+  const result = auctionAdminService.closeActiveAuction({ seasonId, auctionService });
+  return runAndRespond(res, result, () => ({ ok: true }));
 });
 
 router.get('/payout-rules', withAdmin, (req, res) => {
   const seasonId = getActiveSeasonId();
-  return res.json({
-    grand_prix: getEventPayoutRules(seasonId, 'grand_prix'),
-    sprint: getEventPayoutRules(seasonId, 'sprint'),
-    season_bonus: getSeasonBonusRules(seasonId),
-  });
+  return res.json(payoutRulesAdminService.getPayoutRulesForSeason({ seasonId }));
 });
 
 router.patch('/payout-rules', withAdmin, (req, res) => {
   const seasonId = getActiveSeasonId();
-  const { grand_prix, sprint, season_bonus } = req.body || {};
-
-  const updateEventRule = db.prepare(`
-    UPDATE event_payout_rules
-    SET bps = ?, active = ?, label = ?, rank_order = ?
-    WHERE id = ? AND season_id = ?
-  `);
-
-  const updateSeasonBonus = db.prepare(`
-    UPDATE season_bonus_rules
-    SET bps = ?, active = ?, label = ?, rank_order = ?
-    WHERE id = ? AND season_id = ?
-  `);
-
-  db.transaction(() => {
-    [grand_prix, sprint].forEach((rules) => {
-      if (!Array.isArray(rules)) return;
-      rules.forEach((rule) => {
-        updateEventRule.run(
-          parseInt(rule.bps, 10) || 0,
-          rule.active ? 1 : 0,
-          String(rule.label || ''),
-          parseInt(rule.rank_order, 10) || 1,
-          rule.id,
-          seasonId,
-        );
-      });
-    });
-
-    if (Array.isArray(season_bonus)) {
-      season_bonus.forEach((rule) => {
-        updateSeasonBonus.run(
-          parseInt(rule.bps, 10) || 0,
-          rule.active ? 1 : 0,
-          String(rule.label || ''),
-          parseInt(rule.rank_order, 10) || 1,
-          rule.id,
-          seasonId,
-        );
-      });
-    }
-  })();
-
-  recalcSeasonBonuses({ seasonId });
-  req.app.get('io')?.emit('standings:update');
-
-  return res.json({ ok: true });
+  const io = req.app.get('io');
+  const result = payoutRulesAdminService.savePayoutRulesForSeason({
+    seasonId,
+    payload: req.body,
+    io,
+  });
+  return runAndRespond(res, result);
 });
 
 router.post('/results/sync-next', withAdmin, async (req, res) => {
   const seasonId = getActiveSeasonId();
   const provider = req.app.get('resultsProvider');
-  const force = req.query?.force === '1' || req.body?.force === true;
-
-  const result = await syncNextEventFromProvider({
+  const result = await resultsAdminService.syncNextResults({
     seasonId,
     provider,
     io: req.app.get('io'),
-    includeFuture: force,
-    ignoreLock: force,
+    force: parseForceFlag(req),
   });
-
-  if (!result.ok) return res.status(result.status || 500).json({ error: result.error });
-  return res.json({ ok: true, ...result });
+  return runAndRespond(res, result, (payload) => ({ ok: true, ...payload }));
 });
 
 router.post('/results/advance-next', withAdmin, async (req, res) => {
   const seasonId = getActiveSeasonId();
   const provider = req.app.get('resultsProvider');
-
-  const result = await syncNextEventFromProvider({
+  const result = await resultsAdminService.syncNextResults({
     seasonId,
     provider,
     io: req.app.get('io'),
-    includeFuture: true,
-    ignoreLock: true,
+    force: true,
   });
-
-  if (!result.ok) return res.status(result.status || 500).json({ error: result.error });
-  return res.json({ ok: true, ...result });
+  return runAndRespond(res, result, (payload) => ({ ok: true, ...payload }));
 });
 
 router.post('/results/sync-event/:id', withAdmin, async (req, res) => {
   const seasonId = getActiveSeasonId();
-  const eventId = parseInt(req.params.id, 10);
-  const provider = req.app.get('resultsProvider');
-  const force = req.query?.force === '1' || req.body?.force === true;
+  const eventId = parseIdFromParams(res, req.params.id);
+  if (eventId == null) return undefined;
 
-  const result = await syncEventFromProvider({
+  const provider = req.app.get('resultsProvider');
+  const result = await resultsAdminService.syncEventResults({
     seasonId,
     eventId,
     provider,
     io: req.app.get('io'),
-    ignoreLock: force,
+    force: parseForceFlag(req),
   });
-
-  if (!result.ok) return res.status(result.status || 500).json({ error: result.error });
-  return res.json({ ok: true, ...result });
+  return runAndRespond(res, result, (payload) => ({ ok: true, ...payload }));
 });
 
 router.get('/results/event/:id', withAdmin, (req, res) => {
   const seasonId = getActiveSeasonId();
-  const eventId = parseInt(req.params.id, 10);
+  const eventId = parseIdFromParams(res, req.params.id);
+  if (eventId == null) return undefined;
 
-  const event = getEventById(seasonId, eventId);
-  if (!event) return res.status(404).json({ error: 'Event not found' });
-
-  return res.json({
-    event,
-    drivers: getDrivers(seasonId),
-    results: getEventResults(eventId),
-  });
+  const result = resultsAdminService.getEventEditorData({ seasonId, eventId });
+  return runAndRespond(res, result, (payload) => ({
+    event: payload.event,
+    drivers: payload.drivers,
+    results: payload.results,
+  }));
 });
 
 router.patch('/results/event/:id', withAdmin, (req, res) => {
   const seasonId = getActiveSeasonId();
-  const eventId = parseInt(req.params.id, 10);
-  const rows = req.body?.results;
-  const force = req.query?.force === '1' || req.body?.force === true;
+  const eventId = parseIdFromParams(res, req.params.id);
+  if (eventId == null) return undefined;
 
-  if (!Array.isArray(rows) || rows.length === 0) {
-    return res.status(400).json({ error: 'results must be a non-empty array' });
-  }
-
-  const upsertResult = upsertEventResults({
+  const result = resultsAdminService.saveManualResultsAndScore({
     seasonId,
     eventId,
-    rows,
-    manualOverride: true,
+    rows: req.body?.results,
+    force: parseForceFlag(req),
   });
 
-  if (!upsertResult.ok) {
-    return res.status(upsertResult.status || 400).json({ error: upsertResult.error });
-  }
-
-  const scoreResult = scoreEvent({ seasonId, eventId, ignoreLock: force });
-  if (!scoreResult.ok) {
-    return res.status(scoreResult.status || 400).json({ error: scoreResult.error });
+  if (!result.ok) {
+    return res.status(result.status || 400).json({ error: result.error });
   }
 
   req.app.get('io')?.emit('event:scored', { eventId });
   req.app.get('io')?.emit('standings:update');
 
-  return res.json({ ok: true, rowCount: upsertResult.rowCount });
+  return res.json({ ok: true, rowCount: result.rowCount });
 });
 
 router.post('/results/recalc-season-bonuses', withAdmin, (req, res) => {
   const seasonId = getActiveSeasonId();
-  const result = recalcSeasonBonuses({ seasonId });
-  req.app.get('io')?.emit('standings:update');
-  return res.json({ ok: true, ...result });
+  const result = resultsAdminService.recalcSeasonBonusesForSeason({
+    seasonId,
+    io: req.app.get('io'),
+  });
+  return runAndRespond(res, result, (payload) => ({ ok: true, ...payload }));
 });
 
 router.get('/results/season-bonus-payouts', withAdmin, (req, res) => {
   const seasonId = getActiveSeasonId();
-  const rows = db.prepare(`
-    SELECT sbp.id, sbp.category, sbp.amount_cents, sbp.tie_count,
-           p.name as participant_name,
-           d.code as driver_code,
-           d.name as driver_name
-    FROM season_bonus_payouts sbp
-    JOIN participants p ON p.id = sbp.participant_id
-    LEFT JOIN drivers d ON d.id = sbp.driver_id
-    WHERE sbp.season_id = ?
-    ORDER BY sbp.category ASC, sbp.amount_cents DESC, p.name ASC
-  `).all(seasonId);
-
-  const totals = db.prepare(`
-    SELECT category, SUM(amount_cents) as total_cents
-    FROM season_bonus_payouts
-    WHERE season_id = ?
-    GROUP BY category
-    ORDER BY category ASC
-  `).all(seasonId);
-
-  return res.json({ rows, totals });
+  return res.json(resultsAdminService.getSeasonBonusPayouts({ seasonId }));
 });
 
 module.exports = router;
